루아스크립트 실행중 유저데이터 메모리가 해제되면 서버가 뻗는 현상 수정

+++ game/model/session/session.cpp
@@ -36,7 +36,8 @@ session::session(fb::socket<fb::game::session>& socket, listener* listener) :
 
 session::~session()
 {
-    delete this->_master;
+    if(this->_master != nullptr)
+        delete this->_master;
 }
 
 void fb::game::session::send(const fb::ostream& stream, bool encrypt, bool wrap)
@@ -852,6 +853,8 @@ int fb::game::session::builtin_look(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -874,6 +877,8 @@ int fb::game::session::builtin_color(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -896,6 +901,8 @@ int fb::game::session::builtin_money(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -918,6 +925,8 @@ int fb::game::session::builtin_exp(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -940,6 +949,8 @@ int fb::game::session::builtin_base_hp(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -962,6 +973,8 @@ int fb::game::session::builtin_base_mp(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -984,6 +997,8 @@ int fb::game::session::builtin_strength(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -1006,6 +1021,8 @@ int fb::game::session::builtin_dexterity(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -1028,6 +1045,8 @@ int fb::game::session::builtin_intelligence(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -1049,6 +1068,8 @@ int fb::game::session::builtin_item(lua_State* lua)
         return 0;
 
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
     auto item = (fb::game::item*)nullptr;
     if(lua_isnumber(lua, 2))
     {
@@ -1076,6 +1097,8 @@ int fb::game::session::builtin_items(lua_State* lua)
         return 0;
 
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     lua_newtable(lua);
     for(int i = 0; i < item::MAX_SLOT; i++)
@@ -1098,6 +1121,8 @@ int fb::game::session::builtin_item_drop(lua_State* lua)
 
     auto acceptor = lua::env<fb::game::acceptor>("acceptor");
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
     auto index = (uint8_t)thread->tointeger(2);
     auto drop_all = thread->toboolean(3);
 
@@ -1118,6 +1143,8 @@ int fb::game::session::builtin_mkitem(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
     auto name = lua_cp949(lua, 2);
     auto store = argc < 3 ? true : thread->toboolean(3);
 
@@ -1152,12 +1179,16 @@ int fb::game::session::builtin_rmitem(lua_State* lua)
     {
         auto argc = thread->argc();
         auto session = thread->touserdata<fb::game::session>(1);
+        if(session == nullptr)
+            return 0;
         auto index = uint8_t(0);
         auto count = argc < 3 ? 1 : (int)thread->tointeger(3);
 
         if(lua_isuserdata(lua, 2))
         {
             auto item = thread->touserdata<fb::game::item>(2);
+            if(item == nullptr)
+                return 0;
             if(item == nullptr)
                 throw std::exception();
 
@@ -1204,6 +1235,8 @@ int fb::game::session::builtin_state(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
     if(argc == 1)
     {
         thread->pushinteger(session->state());
@@ -1225,6 +1258,8 @@ int fb::game::session::builtin_disguise(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -1251,6 +1286,8 @@ int fb::game::session::builtin_class(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -1295,6 +1332,8 @@ int fb::game::session::builtin_level(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     if(argc == 1)
     {
@@ -1320,6 +1359,8 @@ int fb::game::session::builtin_group(lua_State* lua)
 
     auto argc = thread->argc();
     auto session = thread->touserdata<fb::game::session>(1);
+    if(session == nullptr)
+        return 0;
 
     auto group = session->group();
     if(group == nullptr)
